+++ docs/conf.py
@@ -9,7 +9,8 @@
 project = "GOBLIN lite"
 copyright = "2024, Colm Duffy"
 author = "Colm Duffy"
-release = "0.4.2"
+release = "0.5.0"
+
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration